requirements.txt
psycopg2
django-rest-framework
django
django-rest-swagger
requests
